ETL-Pipeline/requirements.txt
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
scikit-learn==1.7.0
pdfplumber==0.10.2
spacy==3.7.2
pytesseract==0.3.10
pillow==10.1.0
pdf2image==1.17.0
python-dateutil==2.8.2
